BRING-21: Create application infrastructure according to attached document updated
